Fix deployer construction for neuvector-prime-byos
Ensure that the set-up-chart script can retrieve the chart from GitHub
via https by installing the ca-certificates packages.

Update the set-up-chart script to fail if any command fails or if any
of the arguments is not specified.

Run the set-up-chart script from the Dockerfile with bash -vx to show
details on the commands being run when a failure occurs.

Add a minimal serviceAccount entry to the schema.yaml.