WIP Google rancher-prime-byos
Work-in-progress rancher-prime-byos Click-to-Deploy implementation
based upon the github.com/SUSE-Enceladus/rancher-deployer-charts
repo's rancher-cloud chart and the associated cnab/app/run script.

Integrated build and publish of the installer/Dockerfile image as
the Click-to-Deployer primary application image, which is used by
the chart/rancher-prime-byos chart to install the pre-requisites
and then Rancher.

Used neuvector-prime-byos as the basis for this work, and there
are still some lingering references to and NeuVector related text.

TODO:
  - Finish schema.yaml
  - Finish chart/rancher-prime-byos/templates/application.yaml
  - Update README.md appropriately for Rancher.
  - Update app/tester for Rancher