Add ingress-nginx support to the installer.
Add support for downloading the ingres-nginx chart of the Makefile and
update the installer/Dockerfile to unpack it appropriately. Also update
the installer/run scripts to install the helm chart.

Set the default rancherIngressClassName value in the rancher-prime-byos
chart's values.yaml and the schema.yaml to nginx.

TODO: Add the ingress-nginx images to the rancher-prime-byos repo and
plumb in support for specifying those images into the rancher-prime-byos
chart and the installer-run script.