Retrieve ingress-nginx images from gcr.io repo.
Plumb the ingress-nginx images to be pulled from rancher-prime-byos
grr.io repo.

Note that we need also to override the digest entries to be blank
when installing the ingress-nginx helm chart via the installer/run
script because the digests incorporate the image retrieval path and
thus the path is changed.