Make updates to be able to run mpdev verify successfully
Changes Include:
1. Removed the post-delete yaml template files which was causing the
mpdev install tool to fail with the following error:
Exception: Helm hook post-delete is not supported

2. Updated the reference to the application version to use the tag
defined in the global section to fix the exception:
Exception: Application resource must have a valid spec.descriptor.version value

3. Removed the bootstrapPassword from values.yaml, schema.yaml, and app.run so
we rely on rancher to generate the bootstrap password.
With this approach, we do not need to use the google's x-google-marketplace
GENERATED_PASSWORD for rancherBootstrapPassword.

4. Updated the tests in basic-suite.yaml.
Now the test validates that the rancher service endpoint is up
using the internal endpoint. These tests are run as part of the
mpdev verify command.

5. Updated deployerServiceAccount with CUSTOM rbac rules

6. Enabled service account creation by default

7. Few changes to documentation